requirements.txt
# Core ML and Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Explainable AI Libraries
shap>=0.41.0
lime>=0.2.0.1

# MLOps and Experiment Tracking
mlflow>=1.28.0

# API and Web Framework
fastapi>=0.78.0
uvicorn>=0.17.0

# Data Validation and Processing
pydantic>=1.9.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0
pyyaml>=6.0
jsonschema>=4.6.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
imbalanced-learn>=0.12.0